+++ dsmr_api/views/v2.py
@@ -17,7 +17,6 @@ from dsmr_datalogger.models.reading import DsmrReading
 from dsmr_api.filters import DsmrReadingFilter, DayStatisticsFilter, ElectricityConsumptionFilter,\
     GasConsumptionFilter, HourStatisticsFilter
 import dsmr_consumption.services
-import dsmr_backend.services.backend
 
 
 class DsmrReadingViewSet(mixins.CreateModelMixin, mixins.ListModelMixin, viewsets.GenericViewSet):
@@ -134,9 +133,3 @@ class VersionView(APIView):
         return Response({
             'version': '{}.{}.{}'.format(* settings.DSMRREADER_RAW_VERSION[:3]),
         })
-
-
-class StatusView(APIView):
-    """ Returns an overview of all services and their status. Similar to the Status page in the webinterface. """
-    def get(self, request):
-        return Response(dsmr_backend.services.backend.status_info())